Make all fields holding memory sizes size_t
Turns out there are allocations larger than 4 GiB

+++ stats.cpp
@@ -44,10 +44,10 @@ void AllocStats::print() {
   std::cout << "Peak # of buffers:       " << peakBuffers << std::endl;
 }
 
-uint32_t AllocStats::getCurrentMemory() {
+size_t AllocStats::getCurrentMemory() {
   return liveMemory;
 }
 
-uint32_t AllocStats::getCurrentBuffers() {
+size_t AllocStats::getCurrentBuffers() {
   return liveBuffers;
 }
